Master

+++ src/App.vue
@@ -1,8 +1,8 @@
 <template>
   <div class="container mt-5">
     <h1 class="mb-4 text-center">Hungry!<MyImageLoader :image="'hungry.svg'" :className="'logo'" /></h1>
-    <MyTab :tabs="tabsData" :defaultActive="defaultTabActive" @tabHeightChanged="handleTabHeightChanged">
-      <template v-slot:tabContent0>
+    <MyTab :tabs="tabsData" :defaultActive="defaultTabActive" @tabHeightChanged="handleTabHeightChanged" :alturaDisponible="alturaDisponible" >
+      <template v-slot:tabContent0> <!-- Configuration -->
         <MyCard>
           <p>Selecciona qué configuración deseas exportar:</p>
           <MyCheckbox v-for="index in [0,1]" :key="index" :value="index" :label="CONFIG_NAMES[index]" v-model:checkedValues="configs2Export" :group="'configs2Export'" @lastCheckedDeletionAttempt="handleConfigLastCheckedDeletionAttempt" @update:checkedValues="handleUpdateConfigCheckedValues" />
@@ -13,40 +13,49 @@
           <MyFileReader :text="'Importar Configuración'" @fileReaded="handleFileReaded" @fileReadError="handeFileReadError" :maxFileSize="20*1024" :accept="'application/json'"/>
         </MyCard>
       </template>
-      <template v-slot:tabContent1>
+      <template v-slot:tabContent1> <!-- Add new product -->
         <MyCategoriesList :categories="categoriesData" @categorySelected="handleCategorySelected" @categoryLongClick="handleCategoryLongClick" />
         <MySelect :options="supermercados" v-model="supermercado" selectName="supermercado" @select="handleSelect" />
         <MyInput v-model="nuevoProducto" :placeholder="'Añade elementos aquí'" :autofocus="true" />
         <MyButton text="Añadir" @click="handleAddClick" />
-        <MyModal 
-          ref="myModalRef"
-          :title="`Cambiar «${categoria?.text}»`" 
-          message="Introduzca un nuevo nombre para la categoría" 
-          :value="inputText"
-          @inputModalChange="handleInputModalChange"
-        />
       </template>    
       
-      <template v-slot:tabContent2>
+      <template v-slot:tabContent2> <!-- orderBy name -->
         <MyCard :min-height="alturaDisponible" :heightModifier="-152" :borderStyle="'rounded-bottom'">
-          <MyProductList :productList="productsData" orderBy="name" />
+          <MyProductList 
+            :productList="productsData" 
+            orderBy="name"
+            @click:product="handleClickProduct"
+            @longClick:product="handeLongClickProduct"
+          />
         </MyCard>
       </template>
       
-      <template v-slot:tabContent3>
+      <template v-slot:tabContent3> <!-- orderBy categoryId,name -->
         <MyCard :min-height="alturaDisponible" :heightModifier="-152" :borderStyle="'rounded-bottom'">
-          <MyProductList :productList="productsData" orderBy="categoryId" />
+          <MyProductList 
+            :productList="productsData" 
+            orderBy="categoryId"
+            @click:product="handleClickProduct"
+            @longClick:product="handeLongClickProduct"
+          />
         </MyCard>
       </template>
 
-      <template v-slot:tabContent4>
+      <template v-slot:tabContent4> <!-- Shopping List -->
         <MySelect :options="supermercados" selectName="supermercado" @select="handleSelect" :selectedValue="-1" />
         <MyCard :min-height="alturaDisponible" :heightModifier="-202" :borderStyle="'rounded-bottom'">
-          <MyProductList :productList="productsData" orderBy="categoryId" />
+          <MyProductList :productList="productsData" orderBy="categoryId" :selected="true" :canBeDone="true" @click:product="handleShoplistClick" />
         </MyCard>
       </template>
     </MyTab>
   </div>
+  <MyModalSlot ref="myModalSlotRef">
+    <h1 class="text-center">{{ productoSeleccionado?.text }}</h1>
+    <MyButton :text="'Editar Producto'" @click="handleEditarProducto" />
+    <br>
+    <MyButton  class="mt-15" :btn="'danger'" :text="'Eliminar Producto'" @click="handleEliminarProducto" />
+  </MyModalSlot>
 </template>
 
 <script>
@@ -55,15 +64,14 @@ import MyCard from './components/MyCard.vue';
 import MyCategoriesList from './components/MyCategoriesList.vue';
 import MyCheckbox from './components/MyCheckbox.vue';
 import MyTab from './components/MyTab.vue';
-import MyModal from './components/MyModal.vue';
 import MyInput from './components/MyInput.vue';
 import MyButton from './components/MyButton.vue';
 import MySelect from './components/MySelect.vue';
 import MyProductList from './components/MyProductList.vue';
 import MyImageLoader from './components/MyImageLoader.vue';
 import MyFileReader from './components/MyFile.vue';
-
-import { ref } from 'vue';
+import MyModalSlot from './components/MyModalSlot.vue';
+import { ref, createApp } from 'vue';
 import Swal from 'sweetalert2';
 /**logos **/
 const LOCAL_STORAGE_KEYS = ['categoriesData','productsData'];
@@ -72,38 +80,38 @@ const INDEX_PRODUCTOS=1;
 
 const initialData = [
   [
-    {text:'Categoría 1', bgColor:'#d83c3d'},
-    {text:'Categoría 2', bgColor:'#d8993c'},
-    {text:'Categoría 3', bgColor:'#b9d83c'},
-    {text:'Categoría 4', bgColor:'#5bd83c'},
-    {text:'Categoría 5', bgColor:'#3dd87a'},
-    {text:'Categoría 6', bgColor:'#47ffff'},
-    {text:'Categoría 7', bgColor:'#3b7ad7'},
-    {text:'Categoría 8', bgColor:'#5b3cd8'},
-    {text:'Categoría 9', bgColor:'#b83cd8'},
-    {text:'Categoría 10', bgColor:'#d83ba4'},
-    {text:'Categoría 11', bgColor:'#6f1918'},
-    {text:'Categoría 12', bgColor:'#704c1a'},
-    {text:'Categoría 13', bgColor:'#5d6f19'},
-    {text:'Categoría 14', bgColor:'#2b6f18'},
-    {text:'Categoría 15', bgColor:'#1f8448'},
-    {text:'Categoría 16', bgColor:'#196f70'},
-    {text:'Categoría 17', bgColor:'#183c6e'},
-    {text:'Categoría 18', bgColor:'#2c186f'},
-    {text:'Categoría 19', bgColor:'#5e186e'},
-    {text:'Categoría 20', bgColor:'#6e1952'},
+    {id:0,text:'Categoría 1', bgColor:'#d83c3d'},
+    {id:1,text:'Categoría 2', bgColor:'#d8993c'},
+    {id:2,text:'Categoría 3', bgColor:'#b9d83c'},
+    {id:3,text:'Categoría 4', bgColor:'#5bd83c'},
+    {id:4,text:'Categoría 5', bgColor:'#3dd87a'},
+    {id:5,text:'Categoría 6', bgColor:'#47ffff'},
+    {id:6,text:'Categoría 7', bgColor:'#3b7ad7'},
+    {id:7,text:'Categoría 8', bgColor:'#5b3cd8'},
+    {id:8,text:'Categoría 9', bgColor:'#b83cd8'},
+    {id:9,text:'Categoría 10', bgColor:'#d83ba4'},
+    {id:10,text:'Categoría 11', bgColor:'#6f1918'},
+    {id:11,text:'Categoría 12', bgColor:'#704c1a'},
+    {id:12,text:'Categoría 13', bgColor:'#5d6f19'},
+    {id:13,text:'Categoría 14', bgColor:'#2b6f18'},
+    {id:14,text:'Categoría 15', bgColor:'#1f8448'},
+    {id:15,text:'Categoría 16', bgColor:'#196f70'},
+    {id:16,text:'Categoría 17', bgColor:'#183c6e'},
+    {id:17,text:'Categoría 18', bgColor:'#2c186f'},
+    {id:18,text:'Categoría 19', bgColor:'#5e186e'},
+    {id:19,text:'Categoría 20', bgColor:'#6e1952'},
   ],[]];
 
 function downloadJSON(obj, filename = 'hungry.json') {
   const json = JSON.stringify(obj); // Convertir el objeto a formato JSON
-  const blob = new Blob([json], { type:'application/json'});// Crear un objeto Blob con el contenido JSON
-  const link = document.createElement('a');// Crear un enlace <a> para descargar el archivo
+  const blob = new Blob([json], { type:'application/json'}); // Crear un objeto Blob con el contenido JSON
+  const link = document.createElement('a'); // Crear un enlace <a> para descargar el archivo
   link.href = URL.createObjectURL(blob);
-  link.download = filename;// Establecer el nombre de archivo
-  link.style.display= 'none';// Ocultar el enlace y agregarlo al cuerpo del documento
+  link.download = filename; // Establecer el nombre de archivo
+  link.style.display= 'none'; // Ocultar el enlace y agregarlo al cuerpo del documento
   document.body.appendChild(link);
-  link.click();// Simular un clic en el enlace para iniciar la descarga
-  document.body.removeChild(link);// Limpiar el enlace y el objeto Blob
+  link.click(); // Simular un clic en el enlace para iniciar la descarga
+  document.body.removeChild(link); // Limpiar el enlace y el objeto Blob
   URL.revokeObjectURL(link.href);
 }
 
@@ -117,7 +125,7 @@ export default {
     MyFileReader,
     MyImageLoader,
     MyInput,
-    MyModal,
+    MyModalSlot,
     MyProductList,
     MySelect,
     MyTab,
@@ -132,10 +140,14 @@ export default {
       supermercado:'',
       configs2Export:[],
       defaultTabActive:2,
-      alturaDisponible:0
+      alturaDisponible:0,
+      productoSeleccionado:{}
     }
   },
   methods:{
+    findIndex(what,where){return where.findIndex(item => item === what)},
+    findProductIndex(product){this.findIndex(product,this.productsData)},
+    findCategoryIndex(category){this.findIndex(category,this.categoriesData)},
     setAndSave(where,what)
     {
       if(where==LOCAL_STORAGE_KEYS[INDEX_CATEGORIAS]) this.categoriesData=what
@@ -152,21 +164,92 @@ export default {
         })
       downloadJSON({name:'Hungry!',categorias:this.categoriesData,productos:this.productsData})
     },
+    handleEditarProducto(){
+      const vm = createApp(MyCategoriesList, {
+        categories: this.categoriesData // Pasar las propiedades necesarias al componente
+      }).mount(document.createElement('div')); // Montar el componente Vue en un div creado dinámicamente
+      const htmlContent = vm.$el.outerHTML; 
+      Swal.fire({
+        title: 'Seleccionar categoría',
+        html: htmlContent,
+        showCancelButton: true,
+        confirmButtonText: 'Confirmar',
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          // Acción a realizar si se confirma el SweetAlert2
+          // this.handleCategorySelectedEditProduct(vm.$props.selectedCategory); // Llama a la función con la categoría seleccionada
+        }
+      });
+    },
+    handleEliminarProducto(){
+      Swal.fire({
+        title: '¿Estás seguro de que quieres eliminar el producto?',
+        text: 'Esta acción no se puede deshacer',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Sí',
+        cancelButtonText: 'No',
+        customClass: {
+          cancelButton: 'btn btn-success', // Clase CSS para el botón de confirmación (Sí)
+          confirmButton: 'btn btn-danger' // Clase CSS para el botón de cancelación (No)
+        }
+      }).then((result) => {
+        if (result.isConfirmed) {
+          let index=this.findProductIndex(this.productoSeleccionado)
+          this.productsData.splice(index, 1)
+          localStorageService.setItem(LOCAL_STORAGE_KEYS[INDEX_PRODUCTOS],this.productsData)
+          this.productoSeleccionado=null;
+          this.$refs.myModalSlotRef.closeModal();
+
+        }
+      });
+    },
+    handleShoplistClick(product)
+    {
+      let index=this.findProductIndex(product)
+      this.productsData[index].done=!this.productsData[index].done
+    },
+    handleClickProduct(product){
+      let index=this.findProductIndex(product)
+      this.productsData[index].selected=!this.productsData[index].selected
+      this.productsData[index].done=false
+    },
+    handeLongClickProduct(product){
+      this.productoSeleccionado=product;
+      Swal.fire({
+        title: product.text,
+        text: '¿Qué desea hacer?',
+        icon: 'question',
+        showCancelButton: true,
+        confirmButtonText: 'Editar Producto',
+        cancelButtonText: 'Eliminar Producto',
+        customClass: {
+          confirmButton: 'btn btn-success', // Clase CSS para el botón de confirmación (Sí)
+          cancelButton: 'btn btn-danger' // Clase CSS para el botón de cancelación (No)
+        }
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.handleEditarProducto()
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          this.handleEliminarProducto()
+        }
+      });
+
+    },
     handleTabHeightChanged(data){
       this.alturaDisponible=data;
     },
     handleUpdateConfigCheckedValues(data){
       console.log(data);
     },
-    handleConfigLastCheckedDeletionAttempt(data){
-        console.log("swall")
+    handleConfigLastCheckedDeletionAttempt(){
         Swal.fire({
           icon:'error',
           title:'Error',
           html:"Debe seleccionar al menos<br>una configuración a exportar",
           confirmButtonText:'Aceptar'
         })
-      console.log(data);
     },
     handleFileReaded(data){
       if (data.name!="Hungry!")
@@ -194,18 +277,31 @@ export default {
         })
         return false;
     },
-    handleInputModalChange(newText) {
-      this.categoriesData[this.id_categoria].text = newText;
-      localStorageService.setItem(LOCAL_STORAGE_KEYS[INDEX_CATEGORIAS], this.categoriesData);
-    },
     handleCategorySelected(category,index) {
       this.categoria = category;
       this.id_categoria= index;
       this.nuevoProducto = '';
     },
-    handleCategoryLongClick(){
-      this.inputText=this.categoria.text;
-      this.$refs.myModalRef.openModal();
+    handleCategoryLongClick() {
+      Swal.fire({
+        title: `Cambiar «${this.categoria?.text}»`,
+        text: 'Introduzca un nuevo nombre para la categoría',
+        input: 'text',
+        inputValue: this.categoria.text,
+        showCancelButton: true,
+        confirmButtonText: 'Guardar cambios',
+        cancelButtonText: 'Cancelar',
+        customClass: {
+          cancelButton: 'btn btn-success', // Clase CSS para el botón de confirmación (Sí)
+          confirmButton: 'btn btn-info' // Clase CSS para el botón de cancelación (No)
+        }
+      }).then((result) => {
+        if (result.isConfirmed) {
+          const newText = result.value;
+          this.categoriesData[this.id_categoria].text = newText;
+          localStorageService.setItem(LOCAL_STORAGE_KEYS[INDEX_CATEGORIAS], this.categoriesData);
+        }
+      });
     },
     handleAddClick(){
       if (this.nuevoProducto==='')
@@ -228,7 +324,7 @@ export default {
         })
         return false;
       }
-      this.productsData.push({text:this.nuevoProducto,categoria:this.categoriesData[this.id_categoria],id_categoria:this.id_categoria,supermercado:this.supermercados[this.id_supermercado],id_supermercado:this.id_supermercado});
+      this.productsData.push({id:this.productsData.length,text:this.nuevoProducto,categoria:this.categoriesData[this.id_categoria],id_categoria:this.id_categoria,supermercado:this.supermercados[this.id_supermercado],id_supermercado:this.id_supermercado});
       this.nuevoProducto="";
       localStorageService.setItem(LOCAL_STORAGE_KEYS[INDEX_PRODUCTOS],this.productsData);
     },
